This should make it possible to build both a GTK and a Motif (LessTif) based version of oleo.
With this, I'm not supporting a binary which can run GTK or Motif based
on a command line flag, it's a compile time option.
It is still possible to run oleo in either GUI or curses mode though.

I don't know what to do with the po/ stuff at the bottom line of
  configure.in.